# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Hvordan sier jeg pa samisk API
  version: 0.1.0
servers:
  - url: http://localhost:8000/api/v1
paths:
...
  /jobs/{job_id}:
    get:
      summary: Read a queued or completed job
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    VariantCode:
      type: string
      enum:
        - sme
        - smj
        - sma
    JobStatus:
      type: string
      enum:
        - queued
        - running
        - completed
        - failed
...
    PipelineRequest:
      type: object
      required:
        - target_variant
        - include_phonemes
        - include_audio
      properties:
        target_variant:
          $ref: "#/components/schemas/VariantCode"
        target_voice:
          type: string
          nullable: true
        source_text:
          type: string
          nullable: true
        include_phonemes:
          type: boolean
        include_audio:
          type: boolean
    PipelineStage:
      type: object
      required:
        - name
        - status
        - summary
      properties:
        name:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        summary:
          type: string
    PipelineResult:
      type: object
      required:
        - audio_requested
        - audio_available
        - stages
      properties:
        transcript_text:
          type: string
          nullable: true
        translated_text:
          type: string
          nullable: true
        phoneme_text:
          type: string
          nullable: true
        audio_requested:
          type: boolean
        audio_available:
          type: boolean
        audio_url:
          type: string
          nullable: true
        audio_summary:
          type: string
          nullable: true
        stages:
          type: array
          items:
            $ref: "#/components/schemas/PipelineStage"
    JobRecord:
      type: object
      required:
        - id
        - status
        - request
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        request:
          $ref: "#/components/schemas/PipelineRequest"
        result:
          allOf:
            - $ref: "#/components/schemas/PipelineResult"
          nullable: true
        error:
          type: string
          nullable: true
...
    ErrorResponse:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string